Move helper functions to quaternion_transforms.py
