+++ python/PiFinder/pointing_model/pointing_model.py
@@ -1,89 +1,17 @@
 """
 Pointing model functions.
 
-The quaternions use the notation in the form `q_a2b` for a quaternion that
-rotates frame `a` to frame `b` using intrinsic rotation (by post-multiplying the
-quaternions). For example:
+See `quaternion_transforms.py` for explanations and conventions for
+quaternions.
 
-q_a2c = q_a2b * q_a2c
-
-NOTE: 
-
-* All angles are in radians.
-* The quaternions use numpy quaternions and are scalar-first.
-* Some of the constant quaternion terms can be speeded up by not using 
-trigonometric functions.
-* The methods do not normalize the quaternions because this incurs a small 
-computational overhead. Normalization should be done manually as and when 
-necessary.
+NOTE: All angles are in radians.
 """
-from typing import Union  # When updated to Python 3.10+, remove and use new type hints
+
 import numpy as np
 import quaternion
+from typing import Union  # When updated to Python 3.10+, remove and use new type hints
 
-
-def axis_angle2quat(axis, theta):
-    """
-    Convert from axis-angle representation to a quaternion
-
-    INPUTS:
-    axis: (3,) Axis of rotation (doesn't need to be a unit vector)
-    angle: Angle of rotation [rad]
-    """
-    assert(len(axis) == 3, 'axis should be a list or numpy array of length 3.')
-    # Define the vector part of the quaternion
-    v = np.array(axis) / np.linalg.norm(axis) * np.sin(theta / 2)
-
-    return np.quaternion(np.cos(theta / 2), v[0], v[1], v[2])
-
-
-
-def get_q_hor2frame(az, alt):
-    """ 
-    Returns the quaternion to rotate from the horizontal frame to the frame
-    (typically scope) at coordinates (az, alt) for an ideal AltAz mount.
-
-    INPUTS:
-    az: [rad] Azimuth of scope axis
-    alt: [rad] Alt of scope axis
-    """
-    q_az = axis_angle2quat([0, 0, 1], -(az + np.pi / 2))
-    q_alt = axis_angle2quat([1, 0, 0], (np.pi / 2 - alt))
-    return q_az * q_alt
-
-
-def get_azalt_of_q_hor2frame(q_hor2frame):
-    """
-    Returns the (az, alt) pointing of the frame which is defined by the z axis
-    of the q_hor2frame quaternion.
-
-    RETURNS:
-    az: [rad]
-    alt: [rad]
-    """
-    pz = np.quaternion(0, 0, 0, 1)  # Vector z represented as a pure quaternion
-    frame_axis = q_hor2frame * pz * q_hor2frame.conj()  # Returns a pure quaternion along scope axis
-
-    alt = np.pi / 2 - np.arccos(frame_axis.z)
-    az = np.pi - np.arctan2(frame_axis.y, frame_axis.x)
-
-    return az, alt
-
-
-def get_quat_angular_diff(q1, q2):
-    """
-    Calculates the relative rotation between quaternions `q1` and `q2`.
-    Accounts for the double-cover property of quaternions so that if q1 and q2
-    are close, you get small angle d_theta rather than something around 2 * np.pi.
-    """
-    dq = q1.conj() * q2
-    d_theta = 2 * np.arctan2(np.linalg.norm(dq.vec), dq.w)  # atan2 is more robust than using acos
-
-    # Account for double cover where q2 = -q1 gives d_theta = 2 * pi
-    if d_theta > np.pi:
-        d_theta = 2 * np.pi - d_theta
-    
-    return d_theta  # In radians
+import quaternion_transforms as qt
 
 
 class ImuDeadReckoning():
@@ -115,7 +43,7 @@ class ImuDeadReckoning():
     # Dead-reckoning using IMU
     pointing_tracker.update_imu(q_x2imu)
     q_hor2scope = pointing_tracker.get_q_hor2scope()
-    az, alt = get_azalt_of_q_hor2frame(q_hor2scope)
+    az, alt = pointing_tracker.get_cam_azalt()
     """
 
     def __init__(self, screen_direction):
@@ -174,7 +102,7 @@ class ImuDeadReckoning():
         else:
             # Camera frame relative to the horizontal frame where the +y camera
             # frame (i.e. "up") points to zenith:
-            q_hor2cam_up = get_q_hor2frame(solved_cam_az, solved_cam_alt)
+            q_hor2cam_up = qt.get_q_hor2frame(solved_cam_az, solved_cam_alt)
             # Account for camera rotation around the +z camera frame 
             q_cam_rot_z = np.quaternion(np.cos(solved_cam_roll_offset / 2), 
                                         0, 0, np.sin(solved_cam_roll_offset / 2))
@@ -222,7 +150,7 @@ class ImuDeadReckoning():
         indicate if the estimate is from dead-reckoning (True) or from plate
         solving (False).
         """
-        az_cam, alt_cam = get_azalt_of_q_hor2frame(self.q_hor2cam)
+        az_cam, alt_cam = qt.get_azalt_of_q_hor2frame(self.q_hor2cam)
         return az_cam, alt_cam, self.dead_reckoning  # Angles are in radians
 
     def get_scope_azalt(self):
